requirements/common.txt
django==4.0.3
psycopg2==2.9.1
djangorestframework==3.13.1
celery==5.1.2
redis==3.5.3
requests==2.28.1

django-cors-headers==3.13.0

gspread==5.5.0
